z2m_manager/requirements.txt
textual>=0.41.0
PyYAML>=6.0
paho-mqtt>=1.6.1

